Fix state parsing error
State data from the backend service is converted from a full format
(e.g. California) to an abbreviated form for display on the website.
The case statement which implemented this conversion was erroneously
missing `break` statements, causing experiences from "California" to
be represented by the abbreviation for Illinois (IL). This change was
fixed by adding back those missing `break` statements.